+++ src/commands/hardis/org/diagnose/audittrail.ts
@@ -60,6 +60,8 @@ Regular setup actions performed in major orgs are filtered.
   - changedinteractionuseronoff
   - changedmarketinguseroffon
   - changedmarketinguseronoff
+  - changedofflineuseroffon
+  - changedprofileforuserstdtostd
   - changedprofileforuser
   - changedprofileforusercusttostd
   - changedprofileforuserstdtocust
@@ -246,6 +248,7 @@ This command is part of [sfdx-hardis Monitoring](${CONSTANTS.DOC_URL_ROOT}/sales
         'changedmarketinguseroffon',
         'changedmarketinguseronoff',
         'changedManager',
+        "changedofflineuseroffon",
         'changedprofileforuser',
         'changedprofileforusercusttostd',
         'changedprofileforuserstdtocust',
@@ -253,6 +256,7 @@ This command is part of [sfdx-hardis Monitoring](${CONSTANTS.DOC_URL_ROOT}/sales
         'changedroleforuser',
         'changedroleforuserfromnone',
         'changedpassword',
+        "changedprofileforuserstdtostd",
         'changedsfcontentuseroffon',
         'changedUserEmailVerifiedStatusUnverified',
         'changedUserEmailVerifiedStatusVerified',